test(server): split shared lifecycle adapter tests
